move bento directory into src
