+++ .eslintrc.js
@@ -349,8 +349,8 @@ module.exports = {
         'test/**/*.js',
         'extensions/**/test/**/*.js',
         'extensions/**/test-e2e/*.js',
-        'bento/components/**/test/**/*.js',
-        'bento/components/**/test-e2e/*.js',
+        'src/bento/components/**/test/**/*.js',
+        'src/bento/components/**/test-e2e/*.js',
         'ads/**/test/**/*.js',
         'testing/**/*.js',
         'build-system/**/test/*.js',
@@ -460,7 +460,7 @@ module.exports = {
       'files': [
         'src/preact/**',
         'extensions/**/1.0/**',
-        'bento/components/**',
+        'src/bento/components/**',
         '**/storybook/**',
       ],
       'rules': {'local/preact-preferred-props': 2},
