+++ src/bento/components/amp-jwplayer/1.0/amp-jwplayer.js
