+++ src/bento/components/amp-jwplayer/1.0/component.type.js
