+++ src/bento/components/amp-jwplayer/1.0/test/test-amp-jwplayer.js
@@ -5,7 +5,7 @@ import {toggleExperiment} from '#experiments';
 
 import {waitFor} from '#testing/helpers/service';
 
-import * as consent from '../../../../../src/consent';
+import * as consent from '../../../../../consent';
 
 /**
  * @param {Document} doc
