+++ src/bento/components/amp-jwplayer/1.0/test/test-component.js
