+++ tsconfig.json
@@ -7,8 +7,8 @@
     "paths": {
       "#3p/*": ["./3p/*"],
       "#ads/*": ["./ads/*"],
-      "#bento": ["./bento"],
-      "#bento/*": ["./bento/*"],
+      "#bento": ["./src/bento"],
+      "#bento/*": ["./src/bento/*"],
       "#build/*": ["./build/*"],
 
       "#builtins/*": ["./src/builtins/*"],
